finish splitting the queue manager header by responsibility
midi_queue_manager.h still held MIDIQueueLane and MIDIQueueStorage, transport-agnostic
ring/pool primitives with no MIDI knowledge, alongside the two transport managers that
actually belong there. They now live in their own midi_queue_lane.h, completing the split
started in earlier tasks.

Also trimmed two includes that neither traits struct in midi_queue_transports.h actually
uses (midi_queue_definitions.h, model/midi/message.h), and one unused include in
midi_queue_manager.h (definitions_cxx.hpp). Every header in io/midi/ still compiles
standalone; no behavior changed.
